Update release script

+++ scripts/to-release-branch.sh
@@ -7,10 +7,20 @@ fi
 # Save the current branch
 base_branch=`git rev-parse --abbrev-ref HEAD`
 
-rm -rf build
-
 git checkout releases
+if [ $? != "0" ]
+then
+	echo Fail to checkout releases branch.
+	exit
+fi
+
 git merge $base_branch --no-edit
+rm -rf build
+rm -rf typings
+git checkout -- typings/ringcentral.d.ts
+. ./envsetup.sh
+typings install
+npm run build
 rm .gitignore
 git add build
 git add src
